fixing issues with Gemini stricter tool encoding Patch to GROQ an TOGETHER LLMs (and Private) to prevent OpenAI style 1024 character limit
