+++ requirements.txt
@@ -9,7 +9,7 @@ llama-index-llms-together==0.3.1
 llama-index-llms-groq==0.3.1
 llama-index-llms-fireworks==0.3.2
 llama-index-llms-cohere==0.4.1
-llama-index-llms-gemini==0.4.14
+llama-index-llms-google-genai ==0.1.7
 llama-index-llms-bedrock==0.3.8
 llama-index-tools-yahoo-finance==0.3.0
 llama-index-tools-arxiv==0.3.0
